Update echo_server.py
Here's the line that you would need to change to get this to pass both tests at the same time. The `conn` represents the connection to an individual client, and it's the one you want to close when you're done with that client. The `sock` represents the server socket that the server is using to listen for new connections, so you don't want to close that until the program is done running. :)